+++ src/InterruptibleProcessor.php
@@ -15,12 +15,14 @@ class InterruptibleProcessor implements ProcessorInterface
         $this->check = $check;
     }
 
-    public function process(array $stages, $payload)
+    public function process($payload, callable ...$stages)
     {
-        foreach ($stages as $stage) {
-            $payload = call_user_func($stage, $payload);
+        $check = $this->check;
 
-            if (true !== call_user_func($this->check, $payload)) {
+        foreach ($stages as $stage) {
+            $payload = $stage($payload);
+
+            if (true !== $check($payload)) {
                 return $payload;
             }
         }