+++ src/ProcessorInterface.php
@@ -6,10 +6,11 @@ namespace League\Pipeline;
 interface ProcessorInterface
 {
     /**
-     * @param array $stages
+     * Process the payload using multiple stages.
+     *
      * @param mixed $payload
      *
      * @return mixed
      */
-    public function process(array $stages, $payload);
+    public function process($payload, callable ...$stages);
 }